Adding changes related to refactoring datacenter flow into new controller
